Removed 'screen' from @media query
The query was determining if the device was "not a screen with a fine
pointer." I just wanted to the query to check that the screen didn't
have a fine pointer.

The point of this is to get the header to have red texted tabs on my
iPhone, but not computer. This is because my computer (Chrome on
Windows) is shit with smoothing out the red color (probably an OS issue
or something, since my Java programs also experience this as well). My
iPhone, (Safari on an iPhone 8), renders the red text fine, however.

I'm sure that there's a better way to check if the page should use red
or blue text, (maybe checking the color index or some crap, I'm not
sure), but this seems to work on the devices I've tested it with.