Added immediate login reflection to Database::isLoggedIn()
Database::isLoggedIn() can now tell if the user was logged in during the
session of the currently running script. If a login is issued on this
page load, isLoggedIn() detects it, on this page load. Previously, a
subsequent page load was required to reflect logins because cookies
being set by the current script (AKA page) would become visible by the
next script (page).